dashboard/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
paho-mqtt==2.1.0
pydantic==2.10.4
websockets==14.1
python-multipart==0.0.20
